+++ integrations/slack/src/actions/saveThread.ts
@@ -1,7 +1,10 @@
 import { SlackRuntimeContext } from '../configuration';
 import { slackAPI } from '../slack';
 import { ConversationSavedBlock, GeneratedDocSummaryBlock, QueryDisplayBlock } from '../ui';
-import { getInstallationApiClient, getInstallationConfig } from '../utils';
+import { getInstallationApiClient, getInstallationConfig, isSaveThreadMessage } from '../utils';
+
+const RUNTIME_TIME_LIMIT = 30000;
+const APP_ORG_URL = 'https://app.gitbook.com/o/';
 
 /**
  *  Save thread in GitBook as a summary (capture)
@@ -20,6 +23,25 @@ export async function saveThread(
     },
     context: SlackRuntimeContext
 ) {
+    const { accessToken, installation } = await getInstallationConfig(context, teamId);
+
+    // acknowledge the request to the user
+    notifySavingThread({ channel: channelId, thread_ts, userId }, context, accessToken);
+
+    const capturesURL = `${APP_ORG_URL}${installation.target.organization}/captures`;
+    // In some cases, the runtime limit is reached before the capture is finished (e.g large threads)
+    // we notify the user before the runtime limit is reached that the capture
+    const timeoutId = registerNotifyBeforeRuntimeLimit(
+        {
+            channel: channelId,
+            thread_ts,
+            userId,
+        },
+        capturesURL,
+        context,
+        accessToken
+    );
+
     const { capture, followupQuestions } = await createMessageThreadCapture(
         {
             team_id: teamId,
@@ -29,7 +51,8 @@ export async function saveThread(
         context
     );
 
-    const { accessToken } = await getInstallationConfig(context, teamId);
+    // managed to avoid the timeout, clear the timeout
+    clearTimeout(timeoutId);
 
     await slackAPI(
         context,
@@ -39,7 +62,7 @@ export async function saveThread(
             payload: {
                 channel: channelId,
                 blocks: [
-                    ...ConversationSavedBlock(),
+                    ...ConversationSavedBlock(capture.urls.app),
                     ...(capture.output.markdown
                         ? GeneratedDocSummaryBlock({
                               summary: capture.output.markdown,
@@ -57,11 +80,6 @@ export async function saveThread(
         },
         { accessToken }
     );
-
-    // Add custom header(s)
-    return new Response(null, {
-        status: 200,
-    });
 }
 
 function slackTimestampToISOFormat(slackTs) {
@@ -106,6 +124,7 @@ async function createMessageThreadCapture(slackEvent, context: SlackRuntimeConte
         },
         { accessToken }
     );
+
     const { messages = [] } = messageReplies;
 
     // get a permalink to the thread
@@ -135,16 +154,29 @@ async function createMessageThreadCapture(slackEvent, context: SlackRuntimeConte
 
     const capture = startCaptureRes.data;
 
-    const events = messages.map((message) => {
-        const { text, ts, thread_ts } = message;
+    const events = messages
+        .filter((message) => {
+            // ignore messages in thread from any bot (todo: potentially limit to gitbook only)
+            if (message.bot_id) {
+                return false;
+            }
 
-        return {
-            type: 'thread.message',
-            text,
-            timestamp: slackTimestampToISOFormat(ts),
-            ...(ts === thread_ts ? { isFirst: true } : {}),
-        };
-    });
+            if (isSaveThreadMessage(message.text)) {
+                return false;
+            }
+
+            return true;
+        })
+        .map((message) => {
+            const { text, ts, thread_ts } = message;
+
+            return {
+                type: 'thread.message',
+                text,
+                timestamp: slackTimestampToISOFormat(ts),
+                ...(ts === thread_ts ? { isFirst: true } : {}),
+            };
+        });
 
     // add all messages in a thread to a capture
     await installationApiClient.orgs.addEventsToCapture(orgId, capture.id, {
@@ -191,4 +223,83 @@ export async function notifyOnlySupportedThreads(context, team, channel, user) {
         },
         { accessToken }
     );
+}
+
+async function notifySavingThread(
+    {
+        channel,
+        thread_ts,
+        userId,
+    }: {
+        channel: string;
+        thread_ts: string;
+        userId: string;
+    },
+    context: SlackRuntimeContext,
+    accessToken: string
+) {
+    // acknowledge the request to the user
+    await slackAPI(
+        context,
+        {
+            method: 'POST',
+            path: 'chat.postEphemeral', // probably alwasy ephemeral? or otherwise have replies in same thread
+            payload: {
+                channel,
+                text: `_Saving to GitBook..._`,
+                ...(userId ? { user: userId } : {}), // actually shouldn't be optional
+                thread_ts,
+            },
+        },
+        {
+            accessToken,
+        }
+    );
+}
+
+function registerNotifyBeforeRuntimeLimit(
+    {
+        channel,
+        thread_ts,
+        userId,
+    }: {
+        channel: string;
+        thread_ts: string;
+        userId: string;
+    },
+    capturesUrl: string,
+    context: SlackRuntimeContext,
+    accessToken: string
+) {
+    const timeoutId = setTimeout(async () => {
+        // acknowledge the request to the user
+        await slackAPI(
+            context,
+            {
+                method: 'POST',
+                path: 'chat.postMessage',
+                payload: {
+                    channel,
+                    blocks: [
+                        {
+                            type: 'section',
+                            text: {
+                                type: 'mrkdwn',
+                                text: `Thread is being saved. You'll find the capture in <${capturesUrl}|GitBook> shortly.`,
+                            },
+                        },
+                    ],
+                    ...(userId ? { user: userId } : {}), // actually shouldn't be optional
+                    thread_ts,
+                },
+            },
+            {
+                accessToken,
+            }
+        );
+
+        // Set to a value slightly less than the actual runtime limit to notify just before
+    }, RUNTIME_TIME_LIMIT - 1000);
+
+    return timeoutId;
 }