+++ packages/cli/src/publish.ts
@@ -7,6 +7,8 @@ import { buildScriptFromManifest } from './build';
 import { resolveFile } from './manifest';
 import { getAPIClient } from './remote';
 
+const targetAll = api.IntegrationTarget.All;
+
 /**
  * Publish the integration to GitBook.
  * If it already exists, it'll update it.
@@ -22,6 +24,12 @@ export async function publishIntegration(
 
     const api = await getAPIClient(true);
 
+    if (typeof manifest.target === 'string' && manifest.target !== targetAll) {
+        console.log(
+            `ℹ️ Publishing integration with "${manifest.target}" as target for installations. Keep in mind this cannot be changed later.`
+        );
+    }
+
     // Publish the integration.
     const created = await api.integrations.publishIntegration(manifest.name, {
         title: manifest.title,
@@ -36,6 +44,7 @@ export async function publishIntegration(
         configurations: manifest.configurations,
         secrets: manifest.secrets,
         visibility: manifest.visibility,
+        target: manifest.target,
         organization: manifest.organization,
         externalLinks: manifest.externalLinks,
         entities: manifest.entities,